Upgraded to t4v3.3. Also changed the pub type of ir_curve and credit_curve from struct to pointer. More on this later..

+++ jpm_cds/objects/leg.c
@@ -14,7 +14,7 @@
 
 using namespace std;
 
-int calculate_leg( obj_loc_t obj_loc, int id )
+int calculate_leg( obj_loc_t obj_loc, long id )
 {
     // Write me.
